Refactor server for Vercel deployment
