--- .env
@@ -1,2 +1,0 @@
-DATABASE_URL=postgresql://postgres:Polo270392!@db.sijdbbauwwsvzphsokzj.supabase.co:5432/postgres
-UNSPLASH_API_KEY=SEvtaDERuOSpwF9Jcns8DqZkruywPMYLwae9qpz7MX8
