+++ .env.example
@@ -1,0 +1,3 @@
+# Copy this file to `.env` and fill in the secrets before running locally.
+DATABASE_URL=postgresql://USER:PASSWORD@HOST:5432/DATABASE
+UNSPLASH_API_KEY=your-unsplash-api-key
