+++ .gitignore
@@ -1,7 +1,21 @@
+# Dependencies
 node_modules
+
+# Build artifacts
 dist
-.DS_Store
 server/public
-vite.config.ts.*
 *.tar.gz
+
+# OS-specific files
+.DS_Store
+
+# Tooling caches
+vite.config.ts.*
+
+# Environment secrets
 .env
+.env.local
+.env.*
+*.env
+*.env.*
+!.env.example
