+++ server/index.ts
@@ -1,68 +1,26 @@
 import dotenv from "dotenv";
-dotenv.config();
-import express, { type Request, Response, NextFunction } from "express";
-import { registerRoutes } from "./routes";
-import { setupVite, serveStatic, log } from "./vite";
-import { dot } from "node:test/reporters";
+if (process.env.NODE_ENV !== "production") {
+  dotenv.config();
+}
 
-const app = express();
-app.use(express.json());
-app.use(express.urlencoded({ extended: false }));
-
-app.use((req, res, next) => {
-  const start = Date.now();
-  const path = req.path;
-  let capturedJsonResponse: Record<string, any> | undefined = undefined;
-
-  const originalResJson = res.json;
-  res.json = function (bodyJson, ...args) {
-    capturedJsonResponse = bodyJson;
-    return originalResJson.apply(res, [bodyJson, ...args]);
-  };
-
-  res.on("finish", () => {
-    const duration = Date.now() - start;
-    if (path.startsWith("/api")) {
-      let logLine = `${req.method} ${path} ${res.statusCode} in ${duration}ms`;
-      if (capturedJsonResponse) {
-        logLine += ` :: ${JSON.stringify(capturedJsonResponse)}`;
-      }
-
-      if (logLine.length > 80) {
-        logLine = logLine.slice(0, 79) + "…";
-      }
-
-      log(logLine);
-    }
-  });
-
-  next();
-});
+import { createServer } from "http";
+import { createApp } from "./app";
+import { setupVite } from "./vite";
+import { log } from "./logger";
+import { serveStatic } from "./static";
 
 (async () => {
-  const server = await registerRoutes(app);
+  const app = await createApp();
+  const server = createServer(app);
 
-  app.use((err: any, _req: Request, res: Response, _next: NextFunction) => {
-    const status = err.status || err.statusCode || 500;
-    const message = err.message || "Internal Server Error";
-
-    res.status(status).json({ message });
-    throw err;
-  });
-
-  // importantly only setup vite in development and after
-  // setting up all the other routes so the catch-all route
-  // doesn't interfere with the other routes
-  if (app.get("env") === "development") {
+  if (process.env.NODE_ENV === "development") {
     await setupVite(app, server);
   } else {
     serveStatic(app);
   }
 
-  // Serve the app on port 3000 (avoiding Apple's AirPlay on 5000)
-  // this serves both the API and the client.
-  const port = 3000;
- server.listen(port, "0.0.0.0", () => {
-  log(`serving on http://localhost:${port}`);
-});
+  const port = Number(process.env.PORT ?? 3000);
+  server.listen(port, "0.0.0.0", () => {
+    log(`serving on http://localhost:${port}`);
+  });
 })();
