+++ server/vite.ts
@@ -1,29 +1,17 @@
-import express, { type Express } from "express";
+import type { Express } from "express";
 import fs from "fs";
 import path from "path";
 import { createServer as createViteServer, createLogger } from "vite";
-import { type Server } from "http";
+import type { Server } from "http";
 import react from "@vitejs/plugin-react";
 import { nanoid } from "nanoid";
 
-const viteLogger = createLogger();
-
-export function log(message: string, source = "express") {
-  const formattedTime = new Date().toLocaleTimeString("en-US", {
-    hour: "numeric",
-    minute: "2-digit",
-    second: "2-digit",
-    hour12: true,
-  });
-
-  console.log(`${formattedTime} [${source}] ${message}`);
-}
-
 export async function setupVite(app: Express, server: Server) {
+  const viteLogger = createLogger();
   const serverOptions = {
     middlewareMode: true,
     hmr: { server },
-  };
+  } as const;
 
   const vite = await createViteServer({
     configFile: false,
@@ -44,23 +32,18 @@ export async function setupVite(app: Express, server: Server) {
         "@": path.resolve(import.meta.dirname, "..", "client", "src"),
         "@shared": path.resolve(import.meta.dirname, "..", "shared"),
       },
-      // Allow Vite to resolve modules from the root node_modules
       preserveSymlinks: false,
     },
-    // Configure optimized deps to be found in root
     optimizeDeps: {
       entries: [
         path.resolve(import.meta.dirname, "..", "client", "index.html"),
-        path.resolve(import.meta.dirname, "..", "client", "src", "**", "*.{ts,tsx}")
+        path.resolve(import.meta.dirname, "..", "client", "src", "**", "*.{ts,tsx}"),
       ],
     },
-    // Configure module resolution to look in root node_modules
     server: {
       ...serverOptions,
       fs: {
-        allow: [
-          path.resolve(import.meta.dirname, ".."),
-        ],
+        allow: [path.resolve(import.meta.dirname, "..")],
       },
     },
     customLogger: {
@@ -70,7 +53,6 @@ export async function setupVite(app: Express, server: Server) {
         process.exit(1);
       },
     },
-    server: serverOptions,
     appType: "custom",
   });
 
@@ -86,34 +68,17 @@ export async function setupVite(app: Express, server: Server) {
         "index.html",
       );
 
-      // always reload the index.html file from disk incase it changes
       let template = await fs.promises.readFile(clientTemplate, "utf-8");
       template = template.replace(
         `src="/src/main.tsx"`,
         `src="/src/main.tsx?v=${nanoid()}"`,
       );
+
       const page = await vite.transformIndexHtml(url, template);
       res.status(200).set({ "Content-Type": "text/html" }).end(page);
     } catch (e) {
       vite.ssrFixStacktrace(e as Error);
       next(e);
     }
-  });
-}
-
-export function serveStatic(app: Express) {
-  const distPath = path.resolve(import.meta.dirname, "public");
-
-  if (!fs.existsSync(distPath)) {
-    throw new Error(
-      `Could not find the build directory: ${distPath}, make sure to build the client first`,
-    );
-  }
-
-  app.use(express.static(distPath));
-
-  // fall through to index.html if the file doesn't exist
-  app.use("*", (_req, res) => {
-    res.sendFile(path.resolve(distPath, "index.html"));
   });
 }
